resolve msvc level 3 64 bit warnings
darcs-hash:20060919212846-16063-2e6c7c40d1edf44fa28915fd0105838e7d31e7bf.gz

+++ libsofia-sip-ua/sresolv/sres.c
@@ -113,7 +113,7 @@ ssize_t sres_recvfrom(sres_socket_t s, void *b, size_t len, int flags,
   int retval;
   if (fromlen)
     ilen = *fromlen;
-  retval = recvfrom(s, b, len, flags, from, fromlen ? &ilen : NULL);
+  retval = recvfrom(s, b, (int)len, flags, from, fromlen ? &ilen : NULL);
   if (fromlen)
     *fromlen = ilen;
   return (ssize_t)retval;
